Symphony NAND & SSD levels' log patches
I've removed the statements that ask the player to set a size of the program memory & SSD components, as they're not being checked at the moment & are completely meaningless. I also re-worked the instructions in the `symphony_1_nand` level's log to be up to date with the current state of the game.