Temporary Instruction decoder level output size increase & patches to the test assembly of the Pipelining level.
I've temporarily increased the sizes of all of the level outputs of the `symphony_decoder` level, to make the level tests catch incorrect bit size settings of the static indexers, as that doesn't work  because the level output pins cast the incoming values to their bit size, before handing over the value to the test logic. Ideally we'd get pins that don't have a size & don't cast their incoming values at all for this, but there are non such pins right now.

I've also fixed `2` errors in the test assembly for the level `symphony_12_pipelining`.